Move the orchestrator to contrib
